TextBoxs working, camera fix
Finished the coding and implementation for the text boxes. The return
key is used to cycle through text and to activate dialogue when clost to
an NPC. I will need to go over how to use the functions with everyone.
Also fixed the camera to smoothly follow the player and be closer to the
ground.