NPCs, Player, Gamecontrolller and
Created iNPC and NPC classes for generic NPCs.  GameController holds the
lists of objects so only one object is modified at a time. The Player
class is from Ian's combat and one version will be stored with the game
controller, so that instance is modified when needed. Scene Controller
may end up being unique for each scene.